v3: stream_buffer_size flag
Issue #2897

The upcoming merge-sort algorithm will change streaming results
to emit one row at a time. So, we need the ability to consolidate
those into reasonable chunks.

this change adds a stream_buffer_size option to vtgate, which performs
this functionality similar to the way vttablet groups the results
for streaming.